pyproject.toml
[project]
name = "golinks"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Dan Foster", email = "dan@zem.org.uk" }
]
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "pyyaml>=6.0.2",
]

[project.scripts]
golinks = "golinks:main"

[build-system]
requires = ["uv_build>=0.8.15,<0.9.0"]
build-backend = "uv_build"
